sdk/python/requirements.txt
# 数据库驱动
mysql-connector-python==8.0.33
psycopg2-binary==2.9.9

# 其他依赖
python-dateutil==2.8.2

# 可选：定时任务
APScheduler==3.10.4
